refactor(canvas): implement useAddAIQuests hook for streamlined AI quest generation
