+++ src/modules/canvas/Canvas.tsx
@@ -14,11 +14,8 @@ import { useCanvasStore } from "@/stores/canvas/canvas-store";
 import { useNavigate, useSearchParams } from "react-router-dom";
 import { CreateSkillModal } from "./components/CreateSkillModal";
 import type { Skill } from "@/shared/types/skill.type";
-import { initialNodes, isSkillNode } from "./canvas.const";
-import { useGenerateAIContent } from "./hooks/ai/useGenerateAIContent";
-import { useLoadingStore } from "@/stores/loading-store";
-import { useSkillStore } from "@/stores/skill/skill-store";
-import { createQuestNode } from "@/shared/utils/quetes/quest-node";
+import { initialNodes } from "./canvas.const";
+import { useAddAIQuests } from "./hooks/useAddAiQuest";
 
 export const Canvas = () => {
   const [connectionStart, setConnectionStart] = useState<string | null>(null);
@@ -33,9 +30,9 @@ export const Canvas = () => {
   const { cursorMode, setCursorMode } = useCanvasStore();
   const { screenToFlowPosition } = useReactFlow();
 
-  const { generate } = useGenerateAIContent();
-
   useCanvasLoader();
+
+  const { addGeneratedQuests: openAIGenerator } = useAddAIQuests();
 
   const {
     nodes,
@@ -87,66 +84,6 @@ export const Canvas = () => {
     },
     [cursorMode, connectionStart, onConnect, setConnectionStart],
   );
-
-  const setLoading = useLoadingStore((state) => state.setLoading);
-
-  const openAIGenerator = async () => {
-    setLoading(true);
-    try {
-      const quests = await generate();
-      if (!Array.isArray(quests) || quests.length === 0) return;
-
-      const {
-        addNode,
-        markNew,
-        removeNode,
-        setNodes: updateNodes,
-        markModifiedNode,
-        nodes: currentNodes,
-      } = useCanvasStore.getState();
-      const setCurrentSkillId = useSkillStore.getState().setCurrentSkillId;
-
-      // Par défaut, la position du skill est { x: 400, y: 50 }
-      // On place les quêtes en dessous du skill, espacées horizontalement pour éviter le chevauchement
-      const skillNode = currentNodes.find(isSkillNode);
-      const skillX = skillNode ? skillNode.position.x : 400;
-      const skillY = skillNode ? skillNode.position.y : 50;
-
-      // Place quests starting from x = skillX - 100, y = skillY + 150
-      const baseY = skillY + 370;
-      const baseX = skillX - 320;
-      const questSpacingX = 470;
-
-      quests.forEach((quest, index) => {
-        const id = `quest-${crypto.randomUUID()}`;
-        const position = { x: baseX + index * questSpacingX, y: baseY };
-
-        const handleUpdate = (field: string, value: any) => {
-          const updated = useCanvasStore
-            .getState()
-            .nodes.map((n) => (n.id === id ? { ...n, data: { ...n.data, [field]: value } } : n));
-          updateNodes(updated);
-          markModifiedNode(id);
-        };
-
-        const newNode = createQuestNode(id, position, removeNode, handleUpdate, {
-          title: quest.title,
-          xp: quest.xp,
-          difficulty: quest.difficulty,
-          description: quest.description,
-          status: "LOCKED",
-        });
-
-        addNode(newNode);
-        markNew(id);
-        if (skillNode?.id) setCurrentSkillId(skillNode.id);
-      });
-    } catch (error) {
-      return error;
-    } finally {
-      setLoading(false);
-    }
-  };
 
   useEffect(() => {
     setIsModalOpen(createSkillParam);
