+++ src/modules/canvas/components/quest-card-component/QuestDescription.tsx
@@ -1,4 +1,3 @@
-
 type DescriptionTextareaProps = {
   description: string;
   onChange?: (value: string) => void;
